Added rumbling bookshelves & more
Bookshelves now rumble when certain conditions are met, as a way of faking out the player. Pressing Space when the book is not active will no take away two seconds from the timer. 
Final Boss length adjusted to be more in-line with the rest of the game. Top and bottom projectiles also slowed slightly